backend/requirements.txt
flask
python-dotenv
openai
langchain
langgraph
tree_sitter
pathlib
